predict.py
import argparse
import yaml
import torch
import os

# TODO: Import necessary modules
# from src.model.pikans import PIKANs
# from src.data_loader.datasets import load_inference_data
# from src.utils.visualization import save_height_map

def predict(model_path, data_path, output_dir):
    """
    Performs inference using a trained PIKANs model.
    """
    print(f"Loading model from {model_path}")
    # TODO: Load the trained model
    # model = PIKANs(...) # Initialize with correct architecture
    # model.load_state_dict(torch.load(model_path))
    # model.eval()

    print(f"Loading and preprocessing data from {data_path}")
    # TODO: Load and preprocess the inference data
    # input_tensor = load_inference_data(data_path)

    print("Running inference...")
    # with torch.no_grad():
    #     height_map_tensor = model(input_tensor)

    # Post-processing and saving
    output_path = os.path.join(output_dir, "predicted_height_map.png")
    os.makedirs(output_dir, exist_ok=True)
    print(f"Saving result to {output_path}")
    # TODO: Convert tensor to image and save
    # save_height_map(height_map_tensor, output_path)

    print("Inference complete.")


if __name__ == "__main__":
    parser = argparse.ArgumentParser(description="Predict height map using a trained PIKANs model.")
    parser.add_argument('--model_path', type=str, required=True, help="Path to the trained model file (.pth).")
    parser.add_argument('--data_path', type=str, required=True, help="Path to the input data directory.")
    parser.add_argument('--output_dir', type=str, default="outputs/predictions", help="Directory to save the output.")
    args = parser.parse_args()

    predict(args.model_path, args.data_path, args.output_dir)