requirements.txt
torch
torchvision
torchaudio
numpy
opencv-python-headless
pyyaml
matplotlib
scikit-image
tqdm